==> 4_Omicron_HLASupertypes_Analysis.rmd <==
---
title: "R Notebook"
author: paulbuckley
date: 08/03/2022
output: html_document
---


```{r}
library(data.table)
library(dplyr)
library(tidyr)
library(ggpubr)
library(Biostrings)
library(tidyverse)
library(foreach)
library(stringdist)
library(doParallel)

```

## Useful functions

```{r}

# Function to provide a closest match. Used to match HLA Alleles across mixed output styles.
ClosestMatch2 = function(string, stringVector){

  stringVector[amatch(string, stringVector, maxDist=Inf)]

}

```

## This worksheet produces the BA1 supertype binding affinity analysis


# Read in Wuhan binding dataset
- Sample of what the data look like.
```{r}

WUHAN_PEPTIDES_BINDING_SCORES=readRDS("ORIGINAL_PEPTIDES_BINDING_SCORES_V2.rds")
WUHAN_PEPTIDES_BINDING_SCORES %>% head
WUHAN_PEPTIDES_BINDING_SCORES %>% nrow

```


# Read in Omicron mutanta binding dataset
- Sample of what the data look like.
```{r}

OMICRON_PEPTIDES_BINDING_SCORES=readRDS("OMICRON_PEPTIDES_BINDING_SCORES_V2.rds")
OMICRON_PEPTIDES_BINDING_SCORES %>% head
OMICRON_PEPTIDES_BINDING_SCORES %>% nrow

```


# Read in mutations

```{r}
MUTATIONS = readRDS("OMICRON_EPITOPE_MUTATIONS.rds")
MUTATIONS %>% mutate(Length = nchar(Peptide))%>% filter(Protein == 'surface glycoprotein') %>% filter(Length== 9)%>% nrow

```

# join all

```{r}

MUTATIONS %>% inner_join(WUHAN_PEPTIDES_BINDING_SCORES)%>% nrow
MUTATIONS %>% inner_join(OMICRON_PEPTIDES_BINDING_SCORES) %>% nrow

FULL_MUTATIONS_DT=MUTATIONS %>% inner_join(WUHAN_PEPTIDES_BINDING_SCORES) %>% inner_join(OMICRON_PEPTIDES_BINDING_SCORES)

DATA_FOR_ANALYSIS=FULL_MUTATIONS_DT %>% select(Peptide, VariantAlignment, Predicted_Binding, EL_Rank, MT_EL_Rank, Binder,MT_Binder) %>% separate_rows_(cols = c("Predicted_Binding","EL_Rank","MT_EL_Rank","Binder","MT_Binder"),sep=",")

```


```{r}
# Read AB supertypes
AB_SUPERTYPES=fread("HLA_AB_SUPERTYPES.csv") %>% mutate(Allele = gsub("\\*","",Allele))%>%
        mutate(Allele = paste0("HLA-",Allele))%>% dplyr::rename(HLA_Allele = Allele)

```


## spike only

```{r,dpi=300}
# Get DT ready for analysis
DATA_FOR_ANALYSIS=FULL_MUTATIONS_DT %>% filter(Protein == "surface glycoprotein") %>% # filter spike
        select(Peptide, VariantAlignment, Predicted_Binding, nM_41, MT_nM_41, Binder,MT_Binder) %>% # select columns and separate rows
        separate_rows_(cols = c("Predicted_Binding","nM_41","MT_nM_41","Binder","MT_Binder"),sep=",")%>%
        mutate(nM_41 = as.numeric(nM_41), MT_nM_41=as.numeric(MT_nM_41))%>% mutate(Agretopicity = MT_nM_41/nM_41) # clean cols

# Process HLA-C separately. Use simply HLA-C:-X as superytype
# HLA-A
HLA_A_B_AGRETOPICITY=DATA_FOR_ANALYSIS %>% dplyr::rename(HLA_Allele=Predicted_Binding)%>%
        filter(HLA_Allele %in% grep("HLA-A|HLA-B",HLA_Allele,value = T)) %>%
        inner_join(AB_SUPERTYPES) %>%
        filter(! (Binder == "NONBINDER" & MT_Binder == 'NONBINDER'))
# HLA-C
HLA_CDATASET_AGRETOPICITY=DATA_FOR_ANALYSIS %>% dplyr::rename(HLA_Allele=Predicted_Binding)%>% filter(!HLA_Allele %in% grep("HLA-A|HLA-B",HLA_Allele,value = T))%>% mutate(Supertype = str_extract(HLA_Allele,"HLA-C[0-9][0-9]"))%>%
        mutate(Supertype = gsub("HLA\\-","",Supertype))%>%
filter(! (Binder == "NONBINDER" & MT_Binder == 'NONBINDER'))

```

# spike only

```{r,dpi=300, fig.width  = 11, fig.height = 7}
# Create the BA1 Supertype plot: for certain supertypes plot changes in binding affinity between diff pmhc
BA1_SUPERTYPE_BOXPLT=HLA_A_B_AGRETOPICITY %>% rbind(HLA_CDATASET_AGRETOPICITY)%>% filter(Supertype %in% c("A02","A03","B07","C01"))%>%
        dplyr::rename(Wuhan=nM_41, Omicron = MT_nM_41)%>%
        ggpaired(cond1 = "Wuhan", cond2 = "Omicron",line.color = "gray")+facet_wrap(~Supertype,nrow=1)+theme_pubr(base_size = 16)+scale_y_log10()+stat_compare_means(label = "p.signif",label.x.npc = "center")+rotate_x_text(angle=45)+ylab("Log10 Binding Affinity (nM)")

BA1_SUPERTYPE_BOXPLT
saveRDS(BA1_SUPERTYPE_BOXPLT,file="BA1_SUPERTYPE_BOXPLT.rds")

```
