==> 6_MIRA_TCR_ANALYSIS_V2.rmd <==
---
title: "R Notebook"
output: html_document
---


```{r setup, include=FALSE}
library(data.table)
library(dplyr)
library(tidyr)
library(ggpubr)
library(Biostrings)
library(tidyverse)
library(ggsci)
library(doParallel)
library(foreach)
```


```{r}
# Read BA1 mutations in
MUTATIONS = readRDS("OMICRON_EPITOPE_MUTATIONS.rds")
# Read in Fitness Data
FITNESS_DATA=readRDS("LR_IMMUNO_FITNESS_DATA_TRAPP.rds")

```

# Read in MIRA
- create a group for mutated in omi or not
```{r}

MIRA_DATASET = fread(cmd = 'unzip -p MIRA_DATA_FULL.txt.zip')

MIRA_DATASET=MIRA_DATASET %>% mutate(Omicron_mutated = ifelse(Peptide %in% MUTATIONS$Peptide, "OMICRON_MUTATED","WT"))
# Get rid of healthy patients, not of interest here.
MIRA_DATASET=MIRA_DATASET%>% filter(! Cohort %in% "Healthy (No known exposure)")

```

# Filter dataset
- Filter for class I Only and get rid unproductive TCRs

```{r}

MIRA_DATASET=MIRA_DATASET %>% filter(! TCR %in% grep("unproductive",TCR, value = TRUE))
MIRA_DATASET=MIRA_DATASET %>% filter(Data_Class == "ClassI")

```


# What % of patient TCR-Epitope repertoires are mutated?

```{r,dpi=300, fig.width = 13}
# Filter class I
FREQ_EP_REP_MIRA_PLT=MIRA_DATASET %>% filter(Data_Class == "ClassI")%>% group_by(Experiment,Cohort,Omicron_mutated) %>%
  dplyr::summarise(n=n())%>% mutate(Freq=n/sum(n))%>% ungroup%>%  # Calculate frequency of eacgh patient's mutated Ep-TCR repertoire
  tidyr::complete_(cols = c("Experiment", "Omicron_mutated"), fill = list(n=0, Freq=0))%>%
  filter(Omicron_mutated == "OMICRON_MUTATED")%>% arrange(desc(Freq))%>% drop_na()%>% # Plot only mutated set
  ggbarplot(x="Experiment",y="Freq",fill="Cohort",palette = "jco")+
  rotate_x_text()+ylab("Freq. TCR-Epitope Repertoire")+xlab("Subject ID")
FREQ_EP_REP_MIRA_PLT
saveRDS(FREQ_EP_REP_MIRA_PLT,file="FREQ_EP_REP_MIRA_PLT.rds")

# Freq of Ep-TCR repertoire that is mutated
TCR_EPITOPE_REP_ANALYSIS_DT=MIRA_DATASET %>% filter(Data_Class == "ClassI")%>% group_by(Experiment,Cohort,Omicron_mutated) %>% dplyr::summarise(n=n())%>% mutate(Freq=n/sum(n))%>% ungroup%>%
    tidyr::complete_(cols = c("Experiment", "Omicron_mutated"), fill = list(n=0, Freq=0))%>%
    filter(Omicron_mutated == "OMICRON_MUTATED")%>% arrange(desc(Freq))%>% dplyr::rename(TCR_EP_FREQ = Freq)%>% select(Experiment, TCR_EP_FREQ)

```

# What proportion of epitope repertoire are affected?
- simply, what proportion of unique epitopes recognised in patients TCr repertoire, are mutated?

```{r,dpi=300, fig.width = 13}

# Frequency of patient epitopes that are mutated
EP_REP_DT=MIRA_DATASET %>% select(Experiment, Peptide,Cohort, Omicron_mutated)%>% distinct() %>%
  group_by(Experiment,Cohort, Omicron_mutated)%>% dplyr::summarise(n=n())%>% mutate(Freq=n/sum(n))%>%ungroup %>% # calculate frequency
  filter(Omicron_mutated == "OMICRON_MUTATED")%>% arrange(desc(Freq))%>% dplyr::rename(EP_FREQ = Freq)%>%
  select(Experiment, EP_FREQ)

```

# What is the breadth of the different individual's TCR-Epitope repertoires

```{r}
# Look at the size of indivisual TCR-Epitope repertoires
BREADTH_DT=MIRA_DATASET %>% select(Experiment, Cohort, TCR, Peptide) %>% distinct()%>%
  group_by(Experiment,Cohort) %>% dplyr::summarise(SIZE_TCR_EP_REP=n())

```


## Integrate 'pan-HLA' RI scores
- Read in pan HLA RI scores. So for each peptide, we have an RI score across all applicable HLA.
- Anchor this score to each peptide in the TCR repertoires of COVID-19 patients
```{r}
# Read in PAN HLA RI scores
LR_IMMUNO=readRDS("LR_IMMUNO_FITNESS_DATA_PAN_PEPTIDE.rds")%>% select(Peptide,VariantAlignment,LR_Immuno,MT_Binder_Freq)
# Impute an 'LR Immuno' (RI score) of zero for Peptides that are not nutated
NON_MUTATED_LR_IMMUNO = data.table(Peptide=unique(MIRA_DATASET$Peptide),
                                   VariantAlignment="NA",
                                    LR_Immuno=0,
                                    MT_Binder_Freq=1.0)
NON_MUTATED_LR_IMMUNO=NON_MUTATED_LR_IMMUNO %>% filter(!Peptide %in% LR_IMMUNO$Peptide)
LR_IMMUNO=LR_IMMUNO %>% rbind(NON_MUTATED_LR_IMMUNO)

```


```{r}
## Gives count, mean, standard deviation, standard error of the mean, and confidence interval (default 95%).
##   data: a data frame.
##   measurevar: the name of a column that contains the variable to be summariezed
##   groupvars: a vector containing names of columns that contain grouping variables
##   na.rm: a boolean that indicates whether to ignore NA's
##   conf.interval: the percent range of the confidence interval (default is 95%)
summarySE <- function(data=NULL, measurevar, groupvars=NULL, na.rm=FALSE,
                      conf.interval=.95, .drop=TRUE) {
    library(plyr)

    # New version of length which can handle NA's: if na.rm==T, don't count them
    length2 <- function (x, na.rm=FALSE) {
        if (na.rm) sum(!is.na(x))
        else       length(x)
    }

    # This does the summary. For each group's data frame, return a vector with
    # N, mean, and sd
    datac <- ddply(data, groupvars, .drop=.drop,
      .fun = function(xx, col) {
        c(N    = length2(xx[[col]], na.rm=na.rm),
          mean = mean   (xx[[col]], na.rm=na.rm),
          sd   = sd     (xx[[col]], na.rm=na.rm)
        )
      },
      measurevar
    )

    # Rename the "mean" column
    datac <- rename(datac, c("mean" = measurevar))

    datac$se <- datac$sd / sqrt(datac$N)  # Calculate standard error of the mean

    # Confidence interval multiplier for standard error
    # Calculate t-statistic for confidence interval:
    # e.g., if conf.interval is .95, use .975 (above/below), and use df=N-1
    ciMult <- qt(conf.interval/2 + .5, datac$N-1)
    datac$ci <- datac$se * ciMult

    return(datac)
}
```

```{r,dpi=300, fig.width = 18}

BREADTH_DT = BREADTH_DT%>% select(!Cohort)
# List each patient and the cohort from which they belong
COHORT_LIST = MIRA_DATASET %>% select(Experiment,Cohort) %>% distinct()
# Calculate average RI score (LR_Immuno) for eacxh patient, across their entire TCR-Epitope repertoire
ERROR_DATASET <- summarySE(MIRA_DATASET %>% inner_join(LR_IMMUNO), measurevar="LR_Immuno", groupvars=c("Experiment"))
# Combine data with ERROR_DATASET with frequency of mutated TCR-=Ep repertoire, just Epitope repertoire, and sizes of TCR epitope repertoire
TCR_DT=ERROR_DATASET %>% left_join(TCR_EPITOPE_REP_ANALYSIS_DT)%>% left_join(EP_REP_DT)%>% left_join(BREADTH_DT)

```


```{r,dpi=300, fig.width = 18}
# For each patient, visualise their average LR Immuno
LR_IMMUNO_PLT_BA1=TCR_DT %>% inner_join(COHORT_LIST)%>%
  ggplot(aes(x=reorder(Experiment, -LR_Immuno), y=LR_Immuno,fill=Cohort))+geom_bar(stat = "identity")+
    geom_errorbar(aes(ymin=LR_Immuno-se, ymax=LR_Immuno+se), colour="black", width=.5,position=position_dodge(.9))+ # error bars, +/-se
  theme_pubr(base_size = 12)+rotate_x_text()+xlab("Subject ID")+
  scale_fill_jco()
LR_IMMUNO_PLT_BA1
saveRDS(LR_IMMUNO_PLT_BA1,file="LR_IMMUNO_PLT_BA1.rds")

```

## Scatter plots

```{r,dpi=300}
# Scatter plot comparing patient RI score (LR Immuno) vs. TCR Epitope frequency that is mutated in Omicron
TCR_EP_FREQ_PLT_SCATTER =TCR_DT %>% inner_join(COHORT_LIST) %>% ggscatter(x="LR_Immuno",y="TCR_EP_FREQ",color="Cohort", # Points color, shape and size
   conf.int = TRUE, # Add confidence interval
   cor.coef = TRUE, # Add correlation coefficient. see ?stat_cor
   cor.coeff.args = list(method = "pearson", label.sep = "\n"))+theme_pubr(base_size = 16)+ scale_color_jco()+xlab("Subject RI Score")+ylab("Freq. TCRs recognising mutated epitope")
TCR_EP_FREQ_PLT_SCATTER
saveRDS(TCR_EP_FREQ_PLT_SCATTER,file="TCR_EP_FREQ_PLT_SCATTER.rds")
# Scatter plot comparing patient RI score (LR Immuno) vs. Epitope frequency that is mutated in Omicron
EP_FREQ_PLT_SCATTER=TCR_DT %>% inner_join(COHORT_LIST) %>% ggscatter(x="LR_Immuno",y="EP_FREQ",color="Cohort", # Points color, shape and size
   conf.int = TRUE, # Add confidence interval
   cor.coef = TRUE, # Add correlation coefficient. see ?stat_cor
   cor.coeff.args = list(method = "pearson", label.sep = "\n"))+theme_pubr(base_size = 16)+ scale_color_jco()+xlab("Subject RI Score")+ylab("Freq. mutated epitopes")
EP_FREQ_PLT_SCATTER
saveRDS(EP_FREQ_PLT_SCATTER,file="EP_FREQ_PLT_SCATTER.rds")
# Scatter plot comparing patient RI score (LR Immuno) vs. size of total TCR epitope repertoire
SIZE_TCR_EP_REP_PLT_SCATTER=TCR_DT %>% inner_join(COHORT_LIST) %>% ggscatter(x="LR_Immuno",y="SIZE_TCR_EP_REP",color="Cohort", # Points color, shape and size
   conf.int = TRUE, # Add confidence interval
   cor.coef = TRUE, # Add correlation coefficient. see ?stat_cor
   cor.coeff.args = list(method = "pearson", label.sep = "\n"))+theme_pubr(base_size = 16)+scale_y_log10()+ scale_color_jco()+xlab("Subject RI Score")+
  ylab("# TCR-Epitope samples in repertoire")
SIZE_TCR_EP_REP_PLT_SCATTER
saveRDS(SIZE_TCR_EP_REP_PLT_SCATTER,file="SIZE_TCR_EP_REP_PLT_SCATTER.rds")

Age_DT=MIRA_DATASET %>% select(Experiment,Age)%>% distinct()%>% filter(! Age == 'N/A')%>% mutate(Age =as.numeric(Age))
# Scatter plot comparing patient RI score (LR Immuno) vs. patient Age
AGE_PLT_SCATTER=TCR_DT %>% inner_join(COHORT_LIST) %>% inner_join(Age_DT)%>%
    ggscatter(x="LR_Immuno",y="Age",color="Cohort", # Points color, shape and size
   conf.int = TRUE, # Add confidence interval
   cor.coef = TRUE, # Add correlation coefficient. see ?stat_cor
   cor.coeff.args = list(method = "pearson", label.sep = "\n"))+theme_pubr(base_size = 16)+scale_y_log10()+ scale_color_jco()+xlab("Subject RI Score")+
  ylab("Age")
AGE_PLT_SCATTER
saveRDS(AGE_PLT_SCATTER,file="AGE_PLT_SCATTER.rds")

```
